Failing hard and fast for supplemental Unicode planes
The MySQL driver silently truncates text fields that have characters
above 0xffff. Made the meaning of "utf8" encoding more explicit failing
hard for characters outside of the range.

In the future, if extended Unicode support is needed, a new character
set name would need to be added to distinguish it from the current
restricted 'utf8'.