Changed <%URL%> default from 'active' to 'base'
This prevents situations where an old (or worse yet, an intentionally
misconfigured domain) points to the site. A <%URL top%> based construct
in a form might then lead to the form data posting to a wrong
domain. With memory cached template rendering a domain pointed to the
site and then pointed back to a perpetrator could in theory lead to a
data loss.

Use of 'active' should be discouraged and it certainly should not be the
default.