Unicode handling in Web::Mailer
Added tests for Web::Mailer, fixed encoding as appropriate to work with
the byte-oriented MIME::Lite implementation. Thanks to Brian Mielke for
catching this!

Email subject might still be an issue with some clients, coming as 8-bit
text. Should really be q-encoded.

Overall, it might be better to switch from a MIME::Lite to an Email::MIME based
implementation. That should now be easier now that at least some
regression tests are in place.